Fixed bug in velocity computation. Properly handled distortion when computing flow.
There is a bug in the quat to matrix function. Switching to the TF version fixes the issue.
This also removes the need to transpose the rotations.

We weren't handling the distortion properly before when computing flow.
Now, we compute flow in the rectified image, then add the displacement to the initial points.
We then distort the shifted points, and the final flow is the shifted points minus the
original grid. This guarantees that we handled the transition from rectified to distorted
correctly.